backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
PyJWT==2.8.0
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.2
xgboost==2.0.3
joblib==1.3.2
pydantic==2.7.1
fastapi==0.110.2
uvicorn==0.29.0
pyod==1.1.3
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.28.0